+++ dms_variants/codonvarianttable.py
@@ -1934,7 +1934,6 @@ class CodonVariantTable:
                 expand=(0, 0),
             )
             + p9.ylab(mut_desc)
-            + p9.scale_fill_cmap("gnuplot")
         )
 
         if samples is None:
@@ -1999,7 +1998,7 @@ class CodonVariantTable:
         assert "target" not in set(df.columns).union(set(n_variants.columns))
 
         df = (
-            df.groupby(["library", "sample", "mutation_type", "site"])
+            df.groupby(["library", "sample", "mutation_type", "site"], observed=False)
             .aggregate({"count": "sum"})
             .reset_index()
             .merge(n_variants, on=["library", "sample"])